Education and training assignments by detailed occupation

A table of educational and training requirements for many different jobs.

Usage

EDUCATION_ASSIGNMENTS

Format

A six-column table.

Group

The level of the hierarchy an occupation type belongs to.

Major

Which top-level group this occupation belongs to

Minor

Which 2nd-level group this occupation belongs to

Broad

Which 3rd-level group this occupation belongs to

Detailed

The SOC code for this occupation type or group of types.

Title

A verbal description of the occupation type or group.

Source

https://www.bls.gov/emp/ind-occ-matrix/education.xlsx

https://www.bls.gov/emp/tables/education-and-training-by-occupation.htm

See also